fix: Manage attendances by its teacher only

+++ src/atoms/attendances/controller.ts
@@ -3,6 +3,9 @@ import { DeletedResponse, OkResponse } from "../../utils/api/api-response";
 import AttendanceService from "./service";
 import * as validator from "./validator";
 import * as generalValidator from "../../utils/general-validator";
+import { UserWithPermissions } from "../users/types";
+import { ForbiddenError, NotFoundError } from "../../utils/api/api-error";
+import { canTeacherManageAttendance } from "./helpers";
 
 export const getAttendances = async (req: Request, res: Response) => {
   const query = await validator.getAttendances(req);
@@ -19,9 +22,13 @@ export const getAttendances = async (req: Request, res: Response) => {
 export const createAttendance = async (req: Request, res: Response) => {
   const input = await validator.upsertAttendance(req);
 
-  const attendance = await AttendanceService.createAttendance(input);
+  if (await canTeacherManageAttendance(req, input.lectureId)) {
+    const attendance = await AttendanceService.createAttendance(input);
 
-  return new OkResponse(attendance).send(res);
+    return new OkResponse(attendance).send(res);
+  }
+
+  throw new ForbiddenError();
 };
 
 export const deleteAttendance = async (req: Request, res: Response) => {
@@ -29,7 +36,14 @@ export const deleteAttendance = async (req: Request, res: Response) => {
 
   const idNumber = Number(id);
 
-  await AttendanceService.deleteAttendance(idNumber);
+  const attendance = await AttendanceService.getAttendance(idNumber);
+
+  if (!attendance) {
+    throw new NotFoundError();
+  }
+
+  if ((req.user as UserWithPermissions).id === attendance.lecture.teacherId)
+    await AttendanceService.deleteAttendance(idNumber);
 
   return new DeletedResponse("").send(res);
 };
@@ -40,7 +54,14 @@ export const updateAttendance = async (req: Request, res: Response) => {
 
   const idNumber = Number(id);
 
-  const attendance = await AttendanceService.updateAttendance(idNumber, input);
+  if (await canTeacherManageAttendance(req, input.lectureId)) {
+    const attendance = await AttendanceService.updateAttendance(
+      idNumber,
+      input
+    );
 
-  return new OkResponse(attendance).send(res);
+    return new OkResponse(attendance).send(res);
+  }
+
+  throw new ForbiddenError();
 };
